slack/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Slack communication API
slack-bolt

# Environment file handling
python-dotenv